@@ -1,70 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-   <BuildNumber Condition="'$(BuildNumber)' == ''">0</BuildNumber>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-    <DockerComposeProjectPath>..\docker-compose.dcproj</DockerComposeProjectPath>
-    <Configurations>Debug;Release</Configurations>
-    <Company></Company>
-    <Version>1.0.0.$(BuildNumber)</Version>
-    <Product>CryptoWatcher API</Product>
-    <PackageId>CryptoWatcherApi</PackageId>
-    <Authors>cesarbmx</Authors>
-    <UserSecretsId>ecdec8ee-e25b-4673-8b1f-e48a1fd17e67</UserSecretsId>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile></DocumentationFile>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <DocumentationFile></DocumentationFile>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <Content Include="CryptoWatcher.Api.xml">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="8.0.0" />
-    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="6.0.0" />
-    <PackageReference Include="ElmahCore" Version="1.2.3" />
-    <PackageReference Include="FluentValidation.AspNetCore" Version="8.1.2" />
-    <PackageReference Include="Hangfire.AspNetCore" Version="1.7.0-beta1" />
-    <PackageReference Include="Hangfire.Core" Version="1.7.0-beta1" />
-    <PackageReference Include="Hangfire.MemoryStorage.Core" Version="1.4.0" />
-    <PackageReference Include="Hangfire.SqlServer" Version="1.7.0-beta1" />
-    <PackageReference Include="MicroElements.Swashbuckle.FluentValidation" Version="1.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.1.5" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="2.1.4" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="2.1.4" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Log4Net.AspNetCore" Version="2.2.6" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="3.0.0" />
-    <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="3.0.0" />
-    <PackageReference Include="Swashbuckle.AspNetCore.Filters" Version="4.3.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.3" />
-  </ItemGroup>
-
-  <ItemGroup>
+    <ProjectReference Include="..\..\Shared\CesarBmx.Shared.Api\CesarBmx.Shared.Api.csproj" />
     <ProjectReference Include="..\CryptoWatcher.Application\CryptoWatcher.Application.csproj" />
-    <ProjectReference Include="..\CryptoWatcher.BackgroundJobs\CryptoWatcher.BackgroundJobs.csproj" />
-    <ProjectReference Include="..\CryptoWatcher.Domain\CryptoWatcher.Domain.csproj" />
     <ProjectReference Include="..\CryptoWatcher.Persistence\CryptoWatcher.Persistence.csproj" />
-    <ProjectReference Include="..\CryptoWatcher.Shared\CryptoWatcher.Shared.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Update="appsettings.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Update="log4net.config">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
   </ItemGroup>
 
 </Project>
