@@ -1,21 +1,17 @@
-﻿using System.IO;
-using System.Reflection;
-using log4net;
-using log4net.Config;
-using Microsoft.AspNetCore.Hosting;
+﻿using System.Reflection;
+using CesarBmx.Shared.Api.Configuration;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
 namespace CryptoWatcher.Api.Configuration
 {
     public static class Log4NetConfig
     {
-        public static ILoggerFactory ConfigureLog4Net(this ILoggerFactory logger, IHostingEnvironment env)
+        public static ILoggerFactory ConfigureLog4Net(this ILoggerFactory logger, IHostEnvironment environment, IConfiguration configuration)
         {
-            var logRepository = LogManager.GetRepository(Assembly.GetEntryAssembly());
-            XmlConfigurator.Configure(logRepository, new FileInfo("log4net.config"));
-
-            logger.AddLog4Net();
-
+             logger.ConfigureSharedLog4Net(Assembly.GetExecutingAssembly(), environment, configuration);
+            
             return logger;
         }
     }
