@@ -1,12 +1,13 @@
 ﻿using System;
 using System.Linq;
 using AutoMapper;
+using CesarBmx.Shared.Application.Responses;
+using CesarBmx.Shared.Common.Extensions;
 using CesarBmx.Shared.Domain.Entities;
 using CoinMarketCap.Entities;
 using Microsoft.Extensions.DependencyInjection;
 using CryptoWatcher.Application.Responses;
 using CryptoWatcher.Domain.Models;
-using CryptoWatcher.Shared.Extensions;
 using Version = CryptoWatcher.Domain.Models.Version;
 
 namespace CryptoWatcher.Api.Configuration
@@ -17,34 +18,34 @@ namespace CryptoWatcher.Api.Configuration
         {
             services.AddAutoMapper(
 
-                cfg => {
-                    
+                 cfg => {
+
                     // Responses
                     cfg.CreateMap<Version, VersionResponse>();
-                    cfg.CreateMap<Health, HealthResponse>();
-                    cfg.CreateMap<Version, VersionResponse>()
-                        .ForMember(dest => dest.BuildDateTime, opt => opt.MapFrom(src => src.LastBuild.ToString("yyyy/MM/dd HH:mm")))
-                        .ForMember(dest => dest.LastBuildOccurred, opt => opt.MapFrom(src => src.LastBuild.DaysHoursMinutesAndSecondsSinceDate()));
-                    cfg.CreateMap<Health, HealthResponse>();
-                    cfg.CreateMap<Currency, CurrencyResponse>();
-                    cfg.CreateMap<Log, LogResponse>();
-                    cfg.CreateMap<Watcher, WatcherResponse>();
-                    cfg.CreateMap<User, UserResponse>();
-                    cfg.CreateMap<Notification, NotificationResponse>();
-                    cfg.CreateMap<Order, OrderResponse>();
-                    cfg.CreateMap<Indicator, IndicatorResponse>()
-                        .ForMember(dest => dest.Dependencies, opt => opt.MapFrom(src => src.Dependencies.Select(x=>x.DependencyId).ToArray()));
-                    cfg.CreateMap<Line, LineResponse>();
-                    cfg.CreateMap<LineChart, LineChartResponse>();
+                     cfg.CreateMap<Health, HealthResponse>();
+                     cfg.CreateMap<Version, VersionResponse>()
+                         .ForMember(dest => dest.BuildDateTime, opt => opt.MapFrom(src => src.LastBuild.ToString("yyyy/MM/dd HH:mm")))
+                         .ForMember(dest => dest.LastBuildOccurred, opt => opt.MapFrom(src => src.LastBuild.DaysHoursMinutesAndSecondsSinceDate()));
+                     cfg.CreateMap<Health, HealthResponse>();
+                     cfg.CreateMap<Currency, CurrencyResponse>();
+                     cfg.CreateMap<Log, LogResponse>();
+                     cfg.CreateMap<Watcher, WatcherResponse>();
+                     cfg.CreateMap<User, UserResponse>();
+                     cfg.CreateMap<Notification, NotificationResponse>();
+                     cfg.CreateMap<Order, OrderResponse>();
+                     cfg.CreateMap<Indicator, IndicatorResponse>()
+                         .ForMember(dest => dest.Dependencies, opt => opt.MapFrom(src => src.Dependencies.Select(x => x.DependencyId).ToArray()));
+                     cfg.CreateMap<Line, LineResponse>();
+                     cfg.CreateMap<LineChart, LineChartResponse>();
 
                     // Others
                     cfg.CreateMap<TickerEntity, Currency>()
-                        .ForMember(dest => dest.CurrencyId, opt => opt.MapFrom(src => src.Id))
-                        .ForMember(dest => dest.Price, opt => opt.MapFrom(src => Convert.ToDecimal(src.PriceUsd)))
-                        .ForMember(dest => dest.Volume24H, opt => opt.MapFrom(src => Convert.ToDecimal(src.Volume24hUsd)))
-                        .ForMember(dest => dest.MarketCap, opt => opt.MapFrom(src => Convert.ToDecimal(src.MarketCapUsd)))
-                        .ForMember(dest => dest.PercentageChange24H, opt => opt.MapFrom(src => Convert.ToDecimal(src.PercentChange24h)));
-                });
+                         .ForMember(dest => dest.CurrencyId, opt => opt.MapFrom(src => src.Id))
+                         .ForMember(dest => dest.Price, opt => opt.MapFrom(src => Convert.ToDecimal(src.PriceUsd)))
+                         .ForMember(dest => dest.Volume24H, opt => opt.MapFrom(src => Convert.ToDecimal(src.Volume24hUsd)))
+                         .ForMember(dest => dest.MarketCap, opt => opt.MapFrom(src => Convert.ToDecimal(src.MarketCapUsd)))
+                         .ForMember(dest => dest.PercentageChange24H, opt => opt.MapFrom(src => Convert.ToDecimal(src.PercentChange24h)));
+                 }, typeof(Startup));
 
             return services;
         }
