@@ -1,16 +1,18 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using CryptoWatcher.Api.RequestExamples;
+using CesarBmx.Shared.Application.Responses;
 using CryptoWatcher.Application.Requests;
-using CryptoWatcher.Api.ResponseExamples;
 using CryptoWatcher.Application.Responses;
 using CryptoWatcher.Application.Services;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
-using Swashbuckle.AspNetCore.Filters;
+using ErrorResponse = CryptoWatcher.Application.Responses.ErrorResponse;
 
 namespace CryptoWatcher.Api.Controllers
 {
+    [SwaggerResponse(500, Type = typeof(InternalServerErrorResponse))]
+    [SwaggerResponse(401, Type = typeof(UnauthorizedResponse))]
+    [SwaggerResponse(403, Type = typeof(ForbiddenResponse))]
     // ReSharper disable once InconsistentNaming
     public class B_UsersController : Controller
     {
@@ -27,9 +29,6 @@ namespace CryptoWatcher.Api.Controllers
         [HttpGet]
         [Route("users")]
         [SwaggerResponse(200, Type = typeof(List<UserResponse>))]
-        [SwaggerResponse(500, Type = typeof(ErrorResponse))]
-        [SwaggerResponseExample(200, typeof(UserListResponseExample))]
-        [SwaggerResponseExample(500, typeof(InternalServerErrorExample))]
         [SwaggerOperation(Tags = new[] { "Users" }, OperationId = "Users_GetAllUsers")]
         public async Task<IActionResult> GetAllUsers()
         {
@@ -47,10 +46,6 @@ namespace CryptoWatcher.Api.Controllers
         [Route("users/{userId}", Name = "Users_GetUser")]
         [SwaggerResponse(200, Type = typeof(UserResponse))]
         [SwaggerResponse(404, Type = typeof(ErrorResponse))]
-        [SwaggerResponse(500, Type = typeof(ErrorResponse))]
-        [SwaggerResponseExample(200, typeof(UserResponseExample))]
-        [SwaggerResponseExample(404, typeof(NotFoundExample))]
-        [SwaggerResponseExample(500, typeof(InternalServerErrorExample))]
         [SwaggerOperation(Tags = new[] { "Users" }, OperationId = "Users_GetUser")]
         public async Task<IActionResult> GetUser(string userId)
         {
@@ -70,13 +65,6 @@ namespace CryptoWatcher.Api.Controllers
         [SwaggerResponse(400, Type = typeof(ErrorResponse))]
         [SwaggerResponse(409, Type = typeof(ErrorResponse))]
         [SwaggerResponse(422, Type = typeof(ValidationResponse))]
-        [SwaggerResponse(500, Type = typeof(ErrorResponse))]
-        [SwaggerResponseExample(201, typeof(UserResponseExample))]
-        [SwaggerResponseExample(400, typeof(BadRequestExample))]
-        [SwaggerResponseExample(409, typeof(ConflictExample))]
-        [SwaggerResponseExample(422, typeof(ValidationFailedExample))]
-        [SwaggerResponseExample(500, typeof(InternalServerErrorExample))]
-        [SwaggerRequestExample(typeof(AddUserRequest), typeof(AddUserRequestExample))]
         [SwaggerOperation(Tags = new[] { "Users" }, OperationId = "Users_AddUser")]
         public async Task<IActionResult> AddUser([FromBody]AddUserRequest request)
         {
