@@ -1,11 +1,12 @@
 ﻿using CryptoWatcher.Application.FakeRequests;
+using CryptoWatcher.Application.Requests;
 using Swashbuckle.AspNetCore.Filters;
 
 namespace CryptoWatcher.Api.RequestExamples
 {
-    public class AddUserRequestExample : IExamplesProvider
+    public class AddUserRequestExample : IExamplesProvider<AddUserRequest>
     {
-        public object GetExamples()
+        public AddUserRequest GetExamples()
         {
             return AddUserFakeRequest.GetFake_master();
         }
