@@ -1,18 +1,20 @@
-﻿using CryptoWatcher.Application.FakeResponses;
+﻿using System.Collections.Generic;
+using CryptoWatcher.Application.FakeResponses;
+using CryptoWatcher.Application.Responses;
 using Swashbuckle.AspNetCore.Filters;
 
 namespace CryptoWatcher.Api.ResponseExamples
 {
-    public class CurrencyResponseExample : IExamplesProvider
+    public class CurrencyResponseExample : IExamplesProvider<CurrencyResponse>
     {
-        public object GetExamples()
+        public CurrencyResponse GetExamples()
         {
             return CurrencyFakeResponse.GetFake_Bitcoin();
         }
     }
-    public class CurrencyListResponseExample : IExamplesProvider
+    public class CurrencyListResponseExample : IExamplesProvider<List<CurrencyResponse>>
     {
-        public object GetExamples()
+        public List<CurrencyResponse> GetExamples()
         {
             return CurrencyFakeResponse.GetFake_List();
         }
