@@ -1,18 +1,20 @@
-﻿using CryptoWatcher.Application.FakeResponses;
+﻿using System.Collections.Generic;
+using CryptoWatcher.Application.FakeResponses;
+using CryptoWatcher.Application.Responses;
 using Swashbuckle.AspNetCore.Filters;
 
 namespace CryptoWatcher.Api.ResponseExamples
 {
-    public class UserResponseExample : IExamplesProvider
+    public class UserResponseExample : IExamplesProvider<UserResponse>
     {
-        public object GetExamples()
+        public UserResponse GetExamples()
         {
             return UserFakeResponse.GetFake_master();
         }
     }
-    public class UserListResponseExample : IExamplesProvider
+    public class UserListResponseExample : IExamplesProvider<List<UserResponse>>
     {
-        public object GetExamples()
+        public List<UserResponse> GetExamples()
         {
             return UserFakeResponse.GetFake_List();
         }
