@@ -1,18 +1,20 @@
-﻿using CryptoWatcher.Application.FakeResponses;
+﻿using System.Collections.Generic;
+using CryptoWatcher.Application.FakeResponses;
+using CryptoWatcher.Application.Responses;
 using Swashbuckle.AspNetCore.Filters;
 
 namespace CryptoWatcher.Api.ResponseExamples
 {
-    public class WatcherResponseExample : IExamplesProvider
+    public class WatcherResponseExample : IExamplesProvider<WatcherResponse>
     {
-        public object GetExamples()
+        public WatcherResponse GetExamples()
         {
             return WatcherFakeResponse.GetFake_master_Bitcoin_RSI();
         }
     }
-    public class WatcherListResponseExample : IExamplesProvider
+    public class WatcherListResponseExample : IExamplesProvider<List<WatcherResponse>>
     {
-        public object GetExamples()
+        public List<WatcherResponse> GetExamples()
         {
             return WatcherFakeResponse.GetFake_List();
         }
